Update SudokuBruteForce
I didn't too a great deal of testing, but I tried it with one case I knew was true, and one case I knew was false, and it got both right. This is probably going to be it for tonight. As it is I don't think it should be too horrible in terms of efficiency, even though there are a lot of for loops, they don't run too many times each.